+++ ChangeLog.md
@@ -5,7 +5,12 @@ Mirrors change log
 
 ## 4.0.0 / ????-??-??
 
-* Added support for `void` return type syntax from PHP 7.1 - @thekid
+* Added support for nullable type syntax present in PHP 7.1; see the
+  [nullable types RFC](https://wiki.php.net/rfc/nullable_types)
+  (@thekid)
+* Added support for `void` return type syntax present in PHP 7.1, see
+  the [void return type RFC](https://wiki.php.net/rfc/void_return_type)
+  (@thekid)
 * Added support for non-`lang.Generic` instances - @thekid
 * **Heads up: Dropped PHP 5.5 support!** The minimum PHP version now
   required is PHP 5.6, see PR #36